+++ src/lib/utils/stripe-utils.ts
@@ -2,6 +2,8 @@ import Stripe from 'stripe';
 import { addClientDB, updateClientStripeCustomerIdDb } from '@/lib/DB/clients-db';
 import { fetchStripAPIKeyDb, storeWebhookInfoDb, fetchWebhookIdDb, deleteWebhookIdDb } from '@/lib/DB/stripe-db';
 import { getStripeInstance } from '../dal/stripe-dal';
+import { schemaCreateSubscription } from '../zod/schemas';
+import z from 'zod';
 
 let stripe: Stripe | null = null;
 
@@ -57,7 +59,7 @@ export async function findOrCreateStripeCustomerAndLinkClient(
             }
         } catch (error) {
             console.error("Error in updateClientStripeCustomerIdDb or subsequent addClientDB:", error);
-            throw error; 
+            throw error;
         }
     } else {
         console.log("No existing Stripe customer found, creating new one.");
@@ -79,7 +81,7 @@ export async function findOrCreateStripeCustomerAndLinkClient(
             await addClientDB(newClientData, effectiveOrgId as string);
         } catch (error) {
             console.error("Error in creating Stripe customer or addClientDB:", error);
-            throw error; 
+            throw error;
         }
     }
     return customerId;
@@ -168,4 +170,66 @@ export async function createStripeCustomer(customerData: {
     const stripe = await getStripeInstance();
     const customer = await stripe.customers.create(customerData);
     return customer;
+}
+
+
+
+export async function createStripeSubscription(subscriptionData: z.infer<typeof schemaCreateSubscription>) {
+    const { clientEmail, clientName, address, phone_number, price_per_month_grass, serviceType, startDate, endDate, organization_id } = subscriptionData;
+
+    stripe = await getStripeInstanceUnprotected(organization_id)
+    if(!Stripe) throw new Error("No Stripe Intinastance ")
+    let customerId: string;
+
+    // 1. Check for existing Stripe customer or create a new one
+    let customer = await getStripeCustomerByEmail(clientEmail);
+
+    if (!customer) {
+        customer = await createStripeCustomer({
+            email: clientEmail,
+            name: clientName,
+            address: { line1: address },
+            phone: phone_number,
+            metadata: { organization_id: organization_id },
+        });
+        customerId = customer.id;
+    } else {
+        customerId = customer.id;
+    }
+
+    // 2. Create a Stripe Product and Price for the subscription
+    // This assumes a simple product/price model. For more complex scenarios, you might pre-create these.
+    const productName = `Lawn Mowing - ${serviceType} for ${clientName}`;
+    const product = await stripe.products.create({
+        name: productName,
+        type: 'service',
+        metadata: { organization_id: organization_id, serviceType: serviceType },
+    });
+
+    const price = await stripe.prices.create({
+        unit_amount: Math.round(price_per_month_grass * 100), // Convert to cents
+        currency: 'cad',
+        recurring: { interval: 'month' },
+        product: product.id,
+        metadata: { organization_id: organization_id, serviceType: serviceType },
+    });
+
+    // 3. Create the Stripe Subscription
+    const subscription = await stripe.subscriptions.create({
+        customer: customerId,
+        items: [{ price: price.id }],
+        collection_method: 'charge_automatically',
+        days_until_due: 7, // Example: invoice 7 days before renewal
+        metadata: {
+            organization_id: organization_id,
+            clientEmail: clientEmail,
+            serviceType: serviceType,
+            startDate: startDate,
+            endDate: endDate || '' // Store endDate if available
+        },
+    });
+
+    // TODO: Save subscription details to your local database if needed
+
+    return subscription;
 }
